Fix a mistake in the database schema.
Project.id was declared as being a String from a few of the classes that
reference it.

We've been testing exclusively with SQLite (other dbmses don't yet work
for other reasons), and my understanding is that SQLite really only has
one type "string." I'm a little bothered by the fact that this was
working.