Clean up/refactor haas.test_common
The maze of stacked decorators is gone; the new implementation
is much simpler.

As a bonus, The deployment tests seem to pass with the code coverage
tool on -- I don't know what was causing it to fail in the first place.

There's more cleanup to be done -- some of the test-specific decorators
are duplicated, for example -- but this is a good step.