Make migration script "merge" the alembic branches.
The docs say you need to enable both extensions before upgrading anyway,
and I think my thinking when I wrote that was that I'd implement the
script like this (and then I forgot).

This has the advantage that it should actually clean up the old entries
for the revisions in the database, rather than just leaving them as
dead-ends.